requirements.txt
requests
python-dotenv
pandas
plotly==5.22.0
